feat: add RAW INPUT test executable and update gitignore
Add test_raw_input project to exercise the RAW INPUT tracking API:
- test_raw_input.cpp: console-based test program with mouse/keyboard/click/wheel
  event callbacks, command interface (delta/keys/devices/status/stats/help)
- test_raw_input.vcxproj: VS2022 x64 C++17 project
- hid_controller.sln: include new test project
- .gitignore: exclude test build output directory
- logitech.cpp: cast mouse delta to int8_t for HID report compatibility

+++ hid-lib/src/core/logitech.cpp
@@ -52,8 +52,8 @@ void update_mouse_button(LogitechDriver::MouseButton& btn, const MOUSEINPUT& mi)
 
 		// 处理鼠标移动
 		if (mi.dwFlags & MOUSEEVENTF_MOVE) {
-			mouse_report_.x_ = mi.dx;
-			mouse_report_.y_ = mi.dy;
+			mouse_report_.x_ = static_cast<int8_t>(mi.dx);
+			mouse_report_.y_ = static_cast<int8_t>(mi.dy);
 		}
 		else {
 			//鼠标移动状态不需要维护，清零
